Remove scripts folder
One of the two scripts wasn't being used anymore. The other seemed to
throw an error on import.

I actually like keeping the scripts within the components where they're
used, and I don't have any scripts that are so long that keeping them
there would be annoying.

Let's see if this fixes things?